urukul_monitor: move ad9910 and ad9912 monitor class away from one big file
