+++ artiq/coredevice/ad9910.py
@@ -18,8 +18,7 @@ __all__ = [
     "PHASE_MODE_CONTINUOUS", "PHASE_MODE_ABSOLUTE", "PHASE_MODE_TRACKING",
     "RAM_DEST_FTW", "RAM_DEST_POW", "RAM_DEST_ASF", "RAM_DEST_POWASF",
     "RAM_MODE_DIRECTSWITCH", "RAM_MODE_RAMPUP", "RAM_MODE_BIDIR_RAMP",
-    "RAM_MODE_CONT_BIDIR_RAMP", "RAM_MODE_CONT_RAMPUP", "AD9910_REG_FTW",
-    "AD9910_REG_PROFILE0", "AD9910_REG_PROFILE7"
+    "RAM_MODE_CONT_BIDIR_RAMP", "RAM_MODE_CONT_RAMPUP"
 ]
 
 _PHASE_MODE_DEFAULT = -1
@@ -64,10 +63,6 @@ RAM_MODE_CONT_RAMPUP = 4
 
 # Default profile for RAM mode
 _DEFAULT_PROFILE_RAM = 0
-
-AD9910_REG_FTW = _AD9910_REG_FTW
-AD9910_REG_PROFILE0 = _AD9910_REG_PROFILE0
-AD9910_REG_PROFILE7 = _AD9910_REG_PROFILE7
 
 class SyncDataUser:
     def __init__(self, core, sync_delay_seed, io_update_delay):
