+++ artiq/gateware/rtio/phy/urukul_monitor.py
@@ -1,12 +1,10 @@
 from migen import *
 
-from artiq.coredevice.ad9910 import AD9910_REG_FTW, AD9910_REG_PROFILE0, AD9910_REG_PROFILE7
-from artiq.coredevice.ad9912_reg import AD9912_FTW0, AD9912_POW1
-from artiq.coredevice.spi2 import SPI_CONFIG_ADDR, SPI_DATA_ADDR, SPI_END
-from artiq.coredevice.urukul import CS_DDS_CH0, CS_DDS_MULTI, CS_CFG, CFG_IO_UPDATE
+from artiq.coredevice.spi2 import SPI_CONFIG_ADDR
+from artiq.coredevice.urukul import CS_DDS_MULTI, CS_CFG, CFG_IO_UPDATE
 
 
-class _AD9910_AD9912MonitorGeneric(Module):
+class AD9910_AD9912MonitorGeneric(Module):
     def __init__(self, spi_phy, io_update_phy):
         self.spi_phy = spi_phy
         self.io_update_phy = io_update_phy
@@ -39,103 +37,3 @@ class _AD9910_AD9912MonitorGeneric(Module):
         return dest_cfg_reg_and_io_update_bit_set
 
 
-class AD9910Monitor(_AD9910_AD9912MonitorGeneric):
-    def __init__(self, spi_phy, io_update_phy, nchannels=4):
-        data = [Signal(32) for i in range(nchannels)]
-        buffer = [Signal(32) for i in range(nchannels)]
-
-        # Flatten register first, then data
-        self.probes = Array(data)
-        super().__init__(spi_phy, io_update_phy)
-
-        def update_probe_data(i):
-            return If(self.selected(i + CS_DDS_CH0), [
-                data[i].eq(buffer[i])
-            ])
-
-        # 0 -> init, 1 -> start reading
-        state = [Signal(1) for i in range(nchannels)]
-
-        # Bit D0-D4: address, Bit D7: read/write
-        reg = self.current_data[24:29]
-        for i in range(nchannels):
-            self.sync.rio_phy += If(self.selected(i + CS_DDS_CH0) & (self.current_address == SPI_DATA_ADDR), [
-                Case(state[i], {
-                    0: [
-                        # not read
-                        If(~self.current_data[31], [
-                            If((AD9910_REG_PROFILE0 <= reg) & (reg <= AD9910_REG_PROFILE7), [
-                                If(self.length == 8, [
-                                    # assume it is write64, which means the last command in the SPI sequence (SPI_END)
-                                    # will be FTW
-                                    state[i].eq(1)
-                                ])
-                            ]).Elif(reg == AD9910_REG_FTW, [
-                                If((self.length == 24) & (self.flags & SPI_END), [
-                                    # write16
-                                    buffer[i][:16].eq(self.current_data[8:24]),
-                                ]).Elif(self.length == 8, [
-                                    # write32
-                                    state[i].eq(1)
-                                ])
-                            ]),
-
-                        ])
-                    ],
-                    1: [
-                        If(self.flags & SPI_END, [
-                            buffer[i][:32].eq(self.current_data),
-                            state[i].eq(0)
-                        ])
-                    ]
-                })
-            ])
-
-        self.sync.rio_phy += If(self.is_io_update(), [
-            If(self.current_address == SPI_DATA_ADDR, [update_probe_data(i) for i in range(nchannels)]),
-            [state[i].eq(0) for i in range(nchannels)]
-        ])
-
-
-class AD9912Monitor(_AD9910_AD9912MonitorGeneric):
-    def __init__(self, spi_phy, io_update_phy, nchannels=4):
-        data = [Signal(48) for i in range(nchannels)]
-        buffer = [Signal(48) for i in range(nchannels)]
-
-        # Flatten register first, then data
-        self.probes = Array(data)
-        super().__init__(spi_phy, io_update_phy)
-
-        def update_probe_data(i):
-            return If(self.selected(i + CS_DDS_CH0), [
-                data[i].eq(buffer[i])
-            ])
-
-        # 0 -> init, 1 -> start reading
-        state = [Signal(1) for i in range(nchannels)]
-
-        reg = self.current_data[16:29]
-        for i in range(nchannels):
-            self.sync.rio_phy += If(self.selected(i + CS_DDS_CH0) & (self.current_address == SPI_DATA_ADDR), [
-                Case(state[i], {
-                    0: [
-                        # Bits A0-A12: address, Bit D15: read write, total 16 bits, need to pad 16 bits
-                        If((self.length == 16) & (AD9912_FTW0 <= reg) & (reg <= AD9912_POW1) & ~self.current_data[31], [
-                            state[i].eq(1)
-                        ])
-                    ],
-                    1: [
-                        If(self.flags & SPI_END, [
-                            buffer[i][:32].eq(self.current_data),
-                            state[i].eq(0)
-                        ]).Else([
-                            buffer[i][32:].eq(self.current_data[:16])
-                        ])
-                    ]
-                })
-            ])
-
-        self.sync.rio_phy += If(self.is_io_update(), [
-            If(self.current_address == SPI_DATA_ADDR, [update_probe_data(i) for i in range(nchannels)]),
-            [state[i].eq(0) for i in range(nchannels)]
-        ])
